\name{easyrodbcmysql-package}
\alias{easyrodbcmysql-package}
\docType{package}
\title{Simple database interface for MySQL.}
\description{Simple database interface for MySQL. easyrodbcmysql is build on top of RODBC. It provides functions for reading, writing, listing and subsetting tables in MySQL. easyrodbcmysql allow to retrieve, write and subset database's tables in a similar way as data.frame's, using single square brackets db[,,,]. easyrodbcmysql internalise many database operations (otherwise tedious) such as rows and columns subsetting (no SQL needed) and connection and disconnection to / from databases.}
\details{
\tabular{ll}{Package: \tab easyrodbcmysql\cr
Version: \tab 0.7.2\cr
Date: \tab 2012-10-22\cr
Title: \tab Simple database interface for MySQL.\cr
Author: \tab Julien MOEYS <Julien.Moeys@mark.slu.se>\cr
Maintainer: \tab Julien MOEYS <Julien.Moeys@mark.slu.se>\cr
Depends: \tab R (>= 2.15.1), RODBC, easydb\cr
License: \tab AGPL-3\cr
URL: \tab http://easydb.r-forge.r-project.org/\cr}
}
\author{Julien MOEYS <Julien.Moeys@mark.slu.se>}

\keyword{ package }


